Updated client to load new maps

+++ runelite/runescape-api/src/main/java/net/runelite/rs/api/RSScene.java
@@ -56,11 +56,11 @@ public interface RSScene extends Scene
 	@Import("removeGameObject")
 	void removeGameObject(GameObject gameObject);
 
-	byte[][][] getUnderlayIds();
-	void setUnderlayIds(byte[][][] underlayIds);
+	short[][][] getUnderlayIds();
+	void setUnderlayIds(short[][][] underlayIds);
 
-	byte[][][] getOverlayIds();
-	void setOverlayIds(byte[][][] overlayIds);
+	short[][][] getOverlayIds();
+	void setOverlayIds(short[][][] overlayIds);
 
 	byte[][][] getTileShapes();
 	void setTileShapes(byte[][][] tileShapes);
